Tidied the namespace a little.
